Also add support for symbol events in eventConstantCheckingUtils

# Conflicts:
#	package-lock.json
#	package.json
#	src/PhantomCollection/index.js

+++ src/ArbitraryPhantomWrapper.ts
@@ -20,24 +20,16 @@ export {
  * Wraps an arbitrary object with a PhantomCore instance.
  */
 export default class ArbitraryPhantomWrapper extends PhantomCore {
-  /**
-   * @param {any} wrappedValue
-   * @param {Object} options? [default = {}]
-   */
-  constructor(wrappedValue, options = {}) {
+  protected _wrappedValue: unknown = null;
+
+  constructor(wrappedValue: unknown, options = {}) {
     super(options);
 
-    this._wrappedValue = null;
     this._setWrappedValue(wrappedValue);
   }
 
-  /**
-   * Internally sets the wrapped element.
-   *
-   * @param {any} wrappedValue
-   * @return {void}
-   */
-  _setWrappedValue(wrappedValue) {
+  // TODO: [3.0.0] Document
+  _setWrappedValue(wrappedValue: unknown) {
     if (this._wrappedValue) {
       throw new Error("_setWrappedValue cannot be called more than once");
     }
@@ -47,30 +39,25 @@ export default class ArbitraryPhantomWrapper extends PhantomCore {
     }
 
     this._wrappedValue = wrappedValue;
+
+    this.registerCleanupHandler(() => {
+      this._wrappedValue = null;
+    });
   }
 
   /**
    * Retrieves the wrapped object which was specified during the class instance
    * creation.
-   *
-   * @return {any}
    */
   getWrappedValue() {
     return this._wrappedValue;
   }
 
-  /**
-   * @param {Function} destroyHandler? [optional] If defined, will execute
-   * prior to normal destruct operations for this class.
-   * @return {Promise<void>}
-   */
-  async destroy(destroyHandler = () => null) {
+  override async destroy(destroyHandler?: () => void) {
     return super.destroy(async () => {
-      await destroyHandler();
-
-      // IMPORTANT: Setting this AFTER super destroy so that it can potentially
-      // be intercepted by other destruct handlers in extension classes
-      this._wrappedValue = null;
+      if (typeof destroyHandler === "function") {
+        await destroyHandler();
+      }
     });
   }
 }
